Rewritten using std::pair

+++ erohin.vladimir/S1/output_named_list.hpp
@@ -3,12 +3,12 @@
 
 #include <iosfwd>
 #include "list.hpp"
-#include "named_number_list.hpp"
+#include "input_named_list.hpp"
 
 namespace erohin
 {
   std::ostream & printNames(std::ostream & output, const List< named_list > & list);
-  std::ostream & printSums(std::ostream & output, const List< long long > & sum);
+  std::ostream & printSums(std::ostream & output, const List< size_t > & sum);
   std::ostream & printElementsAndSums(std::ostream & output, const List< named_list > & list);
 }
 #endif
